Use the Dokan mountpoint feature
We don't have to use windows drives as mountpoints, any folder can do
which is probably a good thing. Also will lead to things being simpler
since all the implementations can start behaving roughly the same.

Case in point, the dependency on the ConfigFile for the Windows VFS is
now the same than on Mac.

Signed-off-by: Kevin Ottens <kevin.ottens@nextcloud.com>